==> Script_3_multiple_imputation_ACE_constructs.R <==
##Lotte Houtepen last edited 29th May 2018 
# using ALSPAC data extracted on 13th April 2018 and 
# dichotomised with the script 'Script_1_Recode_the_original_variables_to_binary.R' and
# ACE constructs derived with 'Script_2_Calculate_the_ACE_constructs.R'


############################################################ 
#All code used to impute the 19 binary adverse childhood experiences (ACE) variables
#for ALSPAC kids filtered as described in Script_2_Calculate_the_ACE_constructs.R (n=12087) 
#
#To speed up imputation, recommend partly running 
#Code block 4 on a HPC cluster (bluecrystal)
#
#Minimum required input is:
#1. Excel file with description variables 
#   - 'SI_data1_ACE_definitions_overtime.xlsx' 
#2. Data file generated by previous script (Script_2_Calculate_the_ACE_constructs.R), 
#   containing ACE constructs and imputation variables
#   - 'alspacKids_ACE_data_2018'
###############
rm(list=ls())#clear global R environment

##############################
###Adapt according to your analysis
##############################
##HAVE TO CHECK 0-2 (or the script will fail)

#0. Location files
#  change to location where input files are stored:
loc_inp<-'add filepath'
#   change to location output should be stored (make sure this folder exists):
loc_out<-'add filepath'
setwd(loc_out)

#1. id used in previous script, also added to new output here
timeperiod<-c(0,16)
fileid_out<-paste0('alspacKids_ACE_',timeperiod[1],'_',timeperiod[2])

#2. load data
#a. excel file with description variables
adv_description<-data.frame(readxl::read_excel(
  paste0(loc_inp,'SI_data1_ACE_definitions_overtime.xlsx'),
  sheet = 'ACE variables',col_names=TRUE))
#b. Exact spelling for name of the data file for imputation
load(paste0(loc_inp,fileid_out,"_imputation.RData"))

#3 Packages
packages<-c('foreign','readxl','matrixStats','tableone','mice')
source("http://bioconductor.org/biocLite.R")
for(pkg in packages){
  if(!require(pkg,character.only=T)){
    biocLite(pkg,suppressUpdates=TRUE)
    library(pkg,character.only=T)
  }
}
rm(pkg,packages)

#custom function to output everything in a log file
catf <- function(..., file=paste0(loc_out,fileid_out,'_imputation_log.txt'), append=TRUE){
  cat(..., file=file, append=append)
}

######################
#Code block 1: Split gender
######################
catf('\n\n',paste0(Sys.time()),
     'Code block 1: Split gender\n')

##A. Split boy/girl
alspacBoys_ACE_data_2018<-alspacKids_ACE_data_2018[
  grepl('Male',alspacKids_ACE_data_2018[['kz021']]),]
alspacGirls_ACE_data_2018<-alspacKids_ACE_data_2018[
  grepl('Female',alspacKids_ACE_data_2018$kz021),]
##B. Drop unused factor levels
alspacKids_ACE_data_2018<-droplevels(alspacKids_ACE_data_2018)
alspacBoys_ACE_data_2018<-droplevels(alspacBoys_ACE_data_2018)
alspacGirls_ACE_data_2018<-droplevels(alspacGirls_ACE_data_2018)

######################
#Code block 2: Determine variables for imputation
######################
catf('\n\n',paste0(Sys.time()),
     'Code block 2: Determine variables for imputation\n')

imp_standard<-grep('org',names(alspacKids_ACE_data_2018),value=T)

#For imputation to work, auxiliary ACE variables need to have at least 50 people in all levels.
#This needs to be true for boys and girls as these are imputed separately so:
#check for factor variables
table(!sapply(alspacKids_ACE_data_2018[,imp_standard],function(x) is.numeric(x)|is.integer(x)))
imp_factor<-
  imp_standard[!sapply(alspacKids_ACE_data_2018[,imp_standard],function(x) is.numeric(x)|is.integer(x))]
catf('\nThere are',length(imp_standard),'imputation variables,',
     'of which',length(imp_factor),'are categorical.\n')
#i. identify ACE imputation variables with less than 50 people in any factor level:
imp_ACE<-imp_factor[gsub('_org','',imp_factor)%in%adv_description$variable_unique]
imp_ACE_less_n50<-
  imp_ACE[sapply(alspacBoys_ACE_data_2018[,imp_ACE],function(x) any(table(factor(x))<50))|
            sapply(alspacGirls_ACE_data_2018[,imp_ACE],function(x) any(table(factor(x))<50))]
catf('\n',length(imp_ACE_less_n50),'of the',length(imp_ACE),'categorical ACE variables,',
     'have <50 people in all factor levels in either boys or girls.\n')
#ii. Identify which ACE imputation variables, do fullfill criteria for binary version
imp_ACE_use_binary<-imp_ACE_less_n50[
  colSums(alspacBoys_ACE_data_2018[,gsub('_org','',imp_ACE_less_n50)],na.rm=T)>=50&
    colSums(alspacGirls_ACE_data_2018[,gsub('_org','',imp_ACE_less_n50)],na.rm=T)>=50]
#iii. If applicable remove imp_ACE_less_n50 in both original and binary version
rm_imp<-imp_ACE_less_n50[!imp_ACE_less_n50%in%imp_ACE_use_binary]
length(imp_standard)
catf(length(rm_imp),'of these',length(imp_ACE_less_n50),'variables need to be removed')
if(length(rm_imp)>0){imp_standard<-imp_standard[!imp_standard%in%rm_imp]}
length(imp_standard)

#iv. If applicable, make sure binary version is used imp_ACE_use_binary
sum(grepl('_org',imp_standard))
catf(', but',length(imp_ACE_use_binary),'variables can be included in their dichotomised version:\n',
     imp_ACE_use_binary,'\n')
if(length(imp_ACE_use_binary)>0){
  imp_standard[imp_standard%in%imp_ACE_use_binary]<-gsub('_org','',imp_standard[imp_standard%in%imp_ACE_use_binary])}
sum(grepl('_org',imp_standard))

catf('\nSo the total number of imputation variables is',length(imp_standard),'of which',
     sum(grepl('_org',imp_standard)),'are included in their original non-dichotomised format.\n')

#clean up
rm(imp_ACE,imp_ACE_less_n50,imp_ACE_use_binary,imp_factor,rm_imp)

######################
#Code block 3: Create predictor matrix
######################
catf('\n\n',paste0(Sys.time()),
     'Code block 3: Create predictor matrix\n')
ACEmeasures<-#ACEs in correct order
  c("ACEscore_classic", "ACEcat_classic",
    "physical_abuse", "sexual_abuse", "emotional_abuse", "emotional_neglect", 
    "bullying","violence_between_parents", "substance_household", 
    "mental_health_problems_or_suicide","parent_convicted_offence",
    "parental_separation", 
    "ACEscore_extended", "ACEcat_extended",
    "social_class", "financial_difficulties", "neighbourhood", 
    "social_support_child", "social_support_parent", 
    "violence_between_child_and_partner","physical_illness_child", 
    "physical_illness_parent","parent_child_bond")
ACEs<-ACEmeasures[c(3:12,15:23)]
classicACEs<-ACEs[1:10]#ACEs in correct order

ACEmeasures<-
  grep(paste0(ACEmeasures,collapse='|'),names(alspacKids_ACE_data_2018),value=T)
ACEs<-
  grep(paste0(ACEs,collapse='|'),names(alspacKids_ACE_data_2018),value=T)
classicACEs<-
  grep(paste0(classicACEs,collapse='|'),names(alspacKids_ACE_data_2018),value=T)

#Extract default predicator matrix for imputation (everything used, matrix filled with 1)
require(mice)
table(unique(c(imp_standard,ACEmeasures))%in%names(alspacBoys_ACE_data_2018))
ini <- mice(alspacKids_ACE_data_2018[
  ,unique(c(imp_standard,ACEmeasures))], maxit=0, pri=F)
pred_mat <- ini$pred#prediction matrix
meth <- ini$meth#formula to be used
length(meth);dim(pred_mat)
#keep gender variable in but do not use for imputation (want to use in modelling but nothing else)
pred_mat[,"kz021_org"] <- 0
meth["kz021_org"] <- ""
length(meth);dim(pred_mat)
#Want to passively impute calculation sumscores
#BUT not use for imputation others so:
pred_mat[,grep('^ACE',colnames(pred_mat))] <- 0
pred_mat[grep('^ACE',rownames(pred_mat)),] <- 0
length(meth);dim(pred_mat)
#passive imputation for each score
meth[grep('ACEscore_extended_',names(meth),value=T)] <- 
  paste0('~I(',paste0('as.integer(',ACEs,')',collapse = '+'),')')
meth[grep('ACEcat_extended_',names(meth),value=T)] <-
  paste0('~I(',paste0('as.integer(',ACEs,')',collapse = '+'),')')
meth[grep('ACEscore_classic_',names(meth),value=T)] <- 
  paste0('~I(',paste0('as.integer(',classicACEs,')',collapse = '+'),')')
meth[grep('ACEcat_classic_',names(meth),value=T)] <-
  paste0('~I(',paste0('as.integer(',classicACEs,')',collapse = '+'),')')
#post-process to create categorical ACE variables
post <- ini$post
post[grep('ACEcat_extended_',names(post),value=T)] <- 
  "imp[[j]][, i] <- cut(imp[[j]][, i], breaks=c(-1,1,2,5,20),labels=c('0_1low','2lomid','3_5midhi','6_18high'))"
post[grep('ACEcat_classic_',names(post),value=T)] <- 
  "imp[[j]][, i] <- cut(imp[[j]][, i], breaks=c(-1,0,1,3,20),labels=c('0','1','2_3','4+'))"
length(meth);dim(pred_mat)
table(rowSums(pred_mat))#nr predictors per variable
names(which(rowSums(pred_mat)==0))#complete variables

save(meth,pred_mat,alspacBoys_ACE_data_2018,alspacGirls_ACE_data_2018,
     file=paste0(loc_out,fileid_out,'_imputation_data.RData'))

######################
#Code block 4: Imputation (and convergence check) 
#RECOMMEND using high performance cluster (HPC)
#HPC parallel processing code is included, but currently outcommented
######################
catf('\n\n',paste0(Sys.time()),
     'Code block 4: Imputation-RECOMMEND HPC\n')

require(mice)
Sys.time()
#First trial run in boys and girls by creating just one imputed dataset using 30 iterations
imp_boys_test<-mice(alspacBoys_ACE_data_2018[,#Need to match the prediction matrix to the dataset
                                             match(rownames(pred_mat),names(alspacBoys_ACE_data_2018))],
                    m = 1,maxit=30,print=TRUE,method=meth,predictorMatrix=pred_mat,post=post, seed=140817)
Sys.time()

imp_girls_test<-mice(alspacGirls_ACE_data_2018[,#Need to match the prediction matrix to the dataset
                                               match(rownames(pred_mat),names(alspacGirls_ACE_data_2018))],
                     m = 1,maxit=30,print=TRUE,method=meth,pred=pred_mat,post=post,seed=140817)
Sys.time()

#Local model will take very long to finish- 20 minutes per imputation so (10*90)/60 = 30 hours for boys only
imp_boys_1_20<-mice(alspacBoys_ACE_data_20171204[,#Need to match the prediction matrix to the dataset
                                                 match(rownames(pred_mat),names(alspacBoys_ACE_data_20171204))],
                    m = 90,maxit=30,print=TRUE,method=meth,predictorMatrix=pred_mat,post=post,seed=140817)
save(imp_boys,file=paste0(fileid_out,'_imp_boys.RData'))

imp_girls_1_20<-mice(alspacGirls_ACE_data_20171204[,#Need to match the prediction matrix to the dataset
                                                   match(rownames(pred_mat),names(alspacGirls_ACE_data_20171204))],
                     m = 90,maxit=30,print=TRUE,method=meth,predictorMatrix=pred_mat,post=post,seed=140817)
save(imp_girls,file=paste0(fileid_out,'_imp_girls.RData'))

#High Performance Cluster will take ~6h for boys, separate job ~6h for girls
#NOTE: this whole code section is outcommented with '# ' as you do not want to run it locally

# ##Example .sh script
# #!/bin/bash
# #
# #
# #PBS -l nodes=1:ppn=16,walltime=6:00:00
# cd $HOME/ACE #Change this line to the location where you stored the data on your HPC
# R --no-save <alspacBoys_ACE_imputationscript.R> out_alspacBoys_ACE_imputationscript
# 
# ##Example R script named 'alspacBoys_ACE_imputationscript.R'
# #load required packages
# require(mice)
# require(parallel)
# 
# #change to name of file saved with code paste0(fileid_out,'_imputation_data.RData')
# load('alspacKids_ACE_imputation_data.RData')
# 
# #change to generic names
# nhanes=#if the script is used for girls change from Boys to Girls
#   alspacBoys_ACE_data_20171204[,#Need to match the prediction matrix to the dataset
#                                 match(rownames(pred_mat),names(alspacBoys_ACE_data_20171204))]
# 
# #specify nr nodes used
# cores_2_use <- 15
# 
# #Do not change next four lines
# cl <- makeCluster(cores_2_use)
# clusterSetRNGStream(cl, 9956)
# clusterExport(cl,list("nhanes",'pred_mat','meth','post'))
# clusterEvalQ(cl, library(mice))
# #Adapt m, maxit, nnet.MaxNWts based on experience with trial run above
# imp_pars <-
#   parLapply(cl = cl, X = 1:cores_2_use, fun = function(no){
#     mice(nhanes, m = 6, #multiple with cores_2_use to get total nr
#                         #imputated datasets, currently 90 (15*6)
#          maxit=30, printFlag = TRUE,method=meth,predictorMatrix=pred_mat,post=post)
#   })
# stopCluster(cl)
# #to merge the datasets
# #if the script is used for girls change from boys to girls
# imp_boys <- imp_pars[[1]]
# for (n in 2:length(imp_pars)){
#   imp_boys <- 
#     ibind(imp_boys,
#           imp_pars[[n]])
# }
# Sys.time()
# #if the script is used for girls change from boys to girls
# save(imp_boys,
#      #change to paste0(fileid_out,'_imp_boys.RData')
#      file='imp_boys_ACE_0_16.RData')
# 
# #Plots to asses convergence (cannot do this after creating object using as.mids)
# pdf('check_imp_boys_ACE.pdf')
# plot(imp_boys)
# dev.off()
# 
# q()

######################
#Code block 5: Check imputation
######################
catf('\n\n',paste0(Sys.time()),
     'Code block 5: Check imputation\n')

#If you've used HPC script, load in data
load(paste0(loc_inp,'imp_boys_ACE_0_16.RData'))
load(paste0(loc_inp,'imp_girls_ACE_0_16.RData'))

#Plots to assess convergence
pdf(paste0(fileid_out,'_check_imp_boys.pdf'))
plot(imp_boys)
dev.off()
pdf(paste0(fileid_out,'_check_imp_girls.pdf'))
plot(imp_girls)
dev.off()

#Combine genders
imp_all<-rbind(imp_boys,imp_girls)
save(imp_all,file=paste0(loc_inp,'imp_all_ACE_0_16.RData'))

com_all <- complete(imp_all, "long", include = TRUE)
table(is.na(com_all$ACEcat_extended_0_16))
nrow(com_all[com_all$.imp==0,]);colSums(is.na(com_all[com_all$.imp==0,]))

##Describe imputation model:
description_imputation<-
  data.frame('original_var'=names(imp_all$method),
             'Variable'=gsub('_org|_dup','',names(imp_all$method)),
             'Type of variable'=sapply(
               imp_all$data[,match(names(imp_all$method),colnames(imp_all$data))],
               function(x) {cl=class(x);paste0(cl,
                                               ifelse(cl%in%c("character", "factor"),
                                                      paste0(' (',length(levels(factor(x))),')'),''))}),
             'Regression model to predict missing in this variable'=imp_all$method,
             'How variable was entered when used to predict missing in other variables'=
               paste(imp_all$pad$categories[match(names(imp_all$method),
                                                  rownames(imp_all$pad$categories)),2],
                     ifelse(imp_all$pad$categories[match(names(imp_all$method),
                                                         rownames(imp_all$pad$categories)),1]==TRUE,
                            'indicator variables','variable_type')))

description_imputation[[5]]<-as.character(description_imputation[[5]])
description_imputation[[5]][grepl('variable_type',description_imputation[[5]])]<-
  as.character(description_imputation[[3]][grepl('variable_type',description_imputation[[5]])])
description_imputation[[3]]<-gsub('logical','dichotomous',description_imputation[[3]])
description_imputation[[3]]<-gsub('integer','continuous',description_imputation[[3]])
description_imputation[[3]]<-gsub('numeric','continuous',description_imputation[[3]])
description_imputation[[3]]<-gsub('factor','categorical',description_imputation[[3]])
description_imputation[[5]]<-gsub('logical','dichotomous',description_imputation[[5]])
description_imputation[[5]]<-gsub('integer','continuous',description_imputation[[5]])
description_imputation[[5]]<-gsub('numeric','continuous',description_imputation[[5]])
description_imputation[[5]]<-gsub('factor','categorical',description_imputation[[5]])
description_imputation[[4]]<-gsub('logreg','Logistic regression',description_imputation[[4]])
description_imputation[[4]]<-gsub('pmm','Predictive mean matching',description_imputation[[4]])
description_imputation[[4]]<-gsub('polyreg','Polytomous (unordered) regression',description_imputation[[4]])

description_imputation$Variable<-gsub('_',' ',description_imputation$Variable)
description_imputation$Variable<-gsub('0 16yrs','0-16yrs',description_imputation$Variable)
description_imputation$Variable<-gsub('parent child bond','parent-child bond',description_imputation$Variable)
description_imputation$Variable<-gsub('ACEscore classic','ACE-score',description_imputation$Variable)
description_imputation$Variable<-gsub('ACEcat classic','Categorical ACE-score',description_imputation$Variable)
description_imputation$Variable<-gsub('ACEscore extended','Extended ACE-score',description_imputation$Variable)
description_imputation$Variable<-gsub('ACEcat extended','Extended categorical ACE-score',description_imputation$Variable)
description_imputation$Variable<-gsub('kz021','gender',description_imputation$Variable)
description_imputation$Variable<-gsub('sc household 18wgest','Household social class at 18wks gestation',description_imputation$Variable)
description_imputation$Variable<-gsub('c804','Ethnicity child',description_imputation$Variable)
description_imputation$Variable<-gsub('a006','Home ownership mother during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('a525','Marital status mother during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('c645a','Self-reported highest educational level mother',description_imputation$Variable)
description_imputation$Variable<-gsub('c666a','Mother-reported highest educational level partner',description_imputation$Variable)
description_imputation$Variable<-gsub('pb325a','Self-reported highest educational level partner ',description_imputation$Variable)
description_imputation$Variable<-gsub('pb342a','Partner-reported highest educational level mother',description_imputation$Variable)
description_imputation$Variable<-gsub('c520','Difficulty affording food during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('c522','Difficulty affording heating during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('mz028b','Maternal age in years at delivery',description_imputation$Variable)
description_imputation$Variable<-gsub('b032','Parity',description_imputation$Variable)
description_imputation$Variable<-gsub('b370','Maternal depression score (EPDS) at 18 wks gestation',description_imputation$Variable)
description_imputation$Variable<-gsub('c600','Maternal depression score (EPDS) at 32 wks gestation',description_imputation$Variable)
description_imputation$Variable<-gsub('pb260','Partner depression score (EPDS) at 18 wks gestation',description_imputation$Variable)
description_imputation$Variable<-gsub('kz030','Birthweight child in grams',description_imputation$Variable)
description_imputation$Variable<-gsub('kz029','Gestational age in weeks at delivery',description_imputation$Variable)
description_imputation$Variable<-gsub('dw002','Maternal pre-pregnancy weight (Kg)',description_imputation$Variable)
description_imputation$Variable<-gsub('dw042','Maternal pre-pregnancy BMI',description_imputation$Variable)
description_imputation$Variable<-gsub('b122','Mother taking medication for depression during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('b593','Mother became homeless during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('c472','Mother became homeless during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('a600','Mother.s opinion of neighbourhood during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('pb188a','Partner convicted of an offence during pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('b587','Mother divorced since pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('pb177','Partner separated since pregnancy',description_imputation$Variable)
description_imputation$Variable<-gsub('pb098','Partner hard drug use during pregnancy',description_imputation$Variable)

write.csv(description_imputation,file=paste0(loc_inp,'description_imputation.csv'))

##Calculate spread in imputed datasets (cc vs imp mean, imp sd, imp min, imp max):
calc_ACE_prev<-function(i_dataset,imp_long,ACEs,...){
  require(matrixStats)
  colSums(sapply(imp_long[imp_long$.imp%in%i_dataset,ACEs],as.numeric),na.rm=T)/
    colSums(!is.na(imp_long[imp_long$.imp%in%i_dataset,ACEs]))
}
ACEs<-c(#"ACEscore_classic_0_16yrs", "ACEcat_classic_0_16yrs", 
  "physical_abuse_0_16yrs", 
  "sexual_abuse_0_16yrs", "emotional_abuse_0_16yrs", "emotional_neglect_0_16yrs", 
  "bullying_0_16yrs", "violence_between_parents_0_16yrs", "substance_household_0_16yrs", 
  "mental_health_problems_or_suicide_0_16yrs", "parent_convicted_offence_0_16yrs", 
  "parental_separation_0_16yrs", #"ACEscore_extended_0_16yrs", "ACEcat_extended_0_16yrs", 
  "social_class_0_16yrs", "financial_difficulties_0_16yrs", "neighbourhood_0_16yrs", 
  "social_support_child_0_16yrs", "social_support_parent_0_16yrs", 
  "violence_between_child_and_partner_0_16yrs", "physical_illness_child_0_16yrs", 
  "physical_illness_parent_0_16yrs", "parent_child_bond_0_16yrs" 
)
ACE_prevalence_per_dataset<-sapply(as.character(unique(com_all$.imp)),calc_ACE_prev,imp_long=com_all,ACEs=ACEs)
SItab5_compare_imputation_all<-
  data.frame(#'Nr_imputed_datasets'=rep(ncol(ACE_prevalence_per_dataset)-1,nrow(ACE_prevalence_per_dataset)),
    'ACE'=ACEs,
    'Complete cases (n)'=colSums(!is.na(com_all[com_all$.imp%in%0,ACEs])),
    'Missingness (%)'=round(colSums(is.na(com_all[com_all$.imp%in%0,ACEs]))/
                              nrow(com_all[com_all$.imp%in%0,ACEs])*100,2),
    'Prevalence complete cases (%)'=round(ACE_prevalence_per_dataset[,1]*100,2),
    'Prevalence imputed data (mean %)'=
      round(rowMeans(ACE_prevalence_per_dataset[,2:ncol(ACE_prevalence_per_dataset)])*100,2),
    'Fold increase in prevalence from complete cases to mean prevalence in imputed data'=
      round(
        (rowMeans(ACE_prevalence_per_dataset[,2:ncol(ACE_prevalence_per_dataset)])-ACE_prevalence_per_dataset[,1])/
          ACE_prevalence_per_dataset[,1],2),
    'Lowest prevalence across all imputed datasets'=
      round(rowMins(ACE_prevalence_per_dataset[,2:ncol(ACE_prevalence_per_dataset)])*100,2),
    'Highest prevalence across all imputed datasets'=
      round(rowMaxs(ACE_prevalence_per_dataset[,2:ncol(ACE_prevalence_per_dataset)])*100,2)
  )
rownames(SItab5_compare_imputation_all)<-gsub('_',' ',rownames(SItab5_compare_imputation_all))
rownames(SItab5_compare_imputation_all)<-gsub('0 16yrs','0-16yrs',rownames(SItab5_compare_imputation_all))
rownames(SItab5_compare_imputation_all)<-gsub('parent child bond','parent-child bond',
                                              rownames(SItab5_compare_imputation_all))
write.csv(SItab5_compare_imputation_all,file=paste0(loc_out,fileid_out,'SItab5_compare_imputation_all.csv'),row.names = F)

################
#End of script Script_3_multiple_imputation_ACE_constructs.R
################
